+++ NINA.Plugin.TargetScheduler.Test/Planning/PlanMocks.cs
@@ -90,6 +90,7 @@ namespace NINA.Plugin.TargetScheduler.Test.Planning {
             Mock<IExposure> pe = new Mock<IExposure>();
             pe.SetupAllProperties();
             pe.SetupProperty(m => m.DatabaseId, databaseId);
+            pe.SetupProperty(m => m.IsEnabled, true);
             pe.SetupProperty(m => m.FilterName, filterName);
             pe.SetupProperty(m => m.ExposureLength, exposureLength);
             pe.SetupProperty(m => m.Desired, desired);