+++ NINA.Plugin.TargetScheduler/Planning/Entities/PlanningExposure.cs
@@ -10,6 +10,7 @@ namespace NINA.Plugin.TargetScheduler.Planning.Entities {
     public class PlanningExposure : IExposure {
         public string PlanId { get; set; }
         public int DatabaseId { get; set; }
+        public bool IsEnabled { get; set; }
         public string FilterName { get; set; }
         public double ExposureLength { get; set; }
         public int? Gain { get; set; }
@@ -44,6 +45,7 @@ namespace NINA.Plugin.TargetScheduler.Planning.Entities {
         public PlanningExposure(ITarget planTarget, ExposurePlan exposurePlan, ExposureTemplate exposureTemplate) {
             this.PlanId = Guid.NewGuid().ToString();
             this.DatabaseId = exposurePlan.Id;
+            this.IsEnabled = exposurePlan.IsEnabled;
             this.FilterName = exposureTemplate.FilterName;
             this.ExposureLength = exposurePlan.Exposure != -1 ? exposurePlan.Exposure : exposureTemplate.DefaultExposure;
             this.Gain = GetNullableIntValue(exposureTemplate.Gain);
@@ -88,6 +90,7 @@ namespace NINA.Plugin.TargetScheduler.Planning.Entities {
         public override string ToString() {
             StringBuilder sb = new StringBuilder();
             sb.AppendLine($"Id: {PlanId}");
+            sb.AppendLine($"Enabled: {IsEnabled}");
             sb.AppendLine($"FilterName: {FilterName}");
             sb.AppendLine($"ExposureLength: {ExposureLength}");
             sb.AppendLine($"Gain: {Gain}");