+++ NINA.Plugin.TargetScheduler/Planning/PlannerEmulator.cs
@@ -191,6 +191,7 @@ namespace NINA.Plugin.TargetScheduler.Planning {
 
     internal class EmulatedExposure : IExposure {
         public string PlanId { get; set; }
+        public bool IsEnabled { get; set; }
         public string FilterName { get; set; }
         public double ExposureLength { get; set; }
         public int? Gain { get; set; }