+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 5.8.0.0 (beta) - 2025-07-XX
 * Added support for exposure rejection for humidity.
 * Added support for an offset in minutes from twilight acceptable times.
-* EP DISABLE?
+* Added ability to disable an exposure plan.
 
 ## 5.7.1.0 - 2025-06-28
 * Fixed issue with target move and how associated acquired image rows are updated.